# api-docs.json
{"openapi":"3.0.1","info":{"title":"Swagger UI Demo API","description":"Демонстрационный проект для изучения Swagger UI, OpenAPI и AsyncAPI","contact":{"name":"Gulash Team","email":"support@gulash.org"},"license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Пользователи","description":"Управление списком пользователей"}],"paths":{"/api/v1/users":{"get":{"tags":["Пользователи"],"summary":"Получить всех пользователей","description":"Возвращает полный список зарегистрированных пользователей","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["Пользователи"],"summary":"Создать пользователя","description":"Добавляет нового пользователя в систему","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/v1/users/{id}":{"get":{"tags":["Пользователи"],"summary":"Найти пользователя по ID","description":"Возвращает одного пользователя по его идентификатору","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"ID пользователя для поиска","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"404":{"description":"Пользователь не найден"},"200":{"description":"Пользователь найден","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}}},"components":{"schemas":{"UserDto":{"required":["email","name"],"type":"object","properties":{"id":{"type":"integer","description":"Уникальный идентификатор","format":"int64","example":1},"name":{"type":"string","description":"Имя пользователя","example":"Иван"},"email":{"type":"string","description":"Электронная почта","example":"ivan@example.com"}},"description":"Сущность пользователя"}}}}